Source: Make TT entry 10bit instead of 16bit with padding
bench: 3865014

+++ Source/structs.h
@@ -23,8 +23,8 @@ typedef struct spsa {
 } spsa_t;
 
 typedef struct tt_entry {
-  uint32_t hash_key; // "almost" unique chess position identifier
-  int move;
+  uint16_t hash_key; // "almost" unique chess position identifier
+  uint16_t move;
   int16_t score; // score (alpha/beta/PV)
   uint8_t depth; // current search depth
   uint8_t flag;  // flag the type of node (fail-low/fail-high/PV)
